switched to double

+++ src/ImageProcessor/models/FaceResonseDetail.cs
@@ -4,6 +4,6 @@ namespace ImageProcessor.models
     {
         public FaceEmotion emotion { get; set; }
 
-        public int age { get; set; }
+        public double age { get; set; }
     }
 }
